clip: held is a per-button bitmask; enforce the edge cap; validate C edge actions
Host-side half of the second clip adversarial pass:

- ClipStatus.held is a u8 bitmask of clip-held mouse buttons (bit b = button id
  b), matching the firmware and reference client, not a bool. Threaded through the
  crate, C ABI (regenerated header), Python binding, mock, and tests.
- medius_clip_builder_frame reads each action as a byte and validates it (an
  out-of-range MediusAction was undefined behavior) and rejects a frame with more
  than CLIP_EDGES_MAX edges; the crate builder asserts the same cap.

+++ src/mock/mod.rs
@@ -220,7 +220,7 @@ fn clip_status_payload(c: ClipStatus) -> Vec<u8> {
     p.extend_from_slice(&c.underruns.to_le_bytes());
     p.extend_from_slice(&c.overruns.to_le_bytes());
     p.extend_from_slice(&c.seq_gaps.to_le_bytes());
-    p.push(c.held as u8);
+    p.push(c.held);
     p
 }
